% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.add_sanity_check}
\alias{.add_sanity_check}
\title{Adds a sanity check to the list of already performed sanity checks}
\usage{
.add_sanity_check(
  fail_vec,
  description,
  counter_meas,
  data,
  data_name,
  example_size,
  param_name,
  call,
  fail_callback,
  .fail_vec_str,
  .generated_desc
)
}
\arguments{
\item{fail_vec}{see \link{add_sanity_check}}

\item{description}{see \link{add_sanity_check}}

\item{counter_meas}{see \link{add_sanity_check}}

\item{data}{see \link{add_sanity_check}}

\item{data_name}{see \link{add_sanity_check}}

\item{example_size}{see \link{add_sanity_check}}

\item{param_name}{see \link{add_sanity_check}}

\item{call}{see \link{add_sanity_check}}

\item{fail_callback}{see \link{add_sanity_check}}

\item{.fail_vec_str}{should capture what was used originally for
\code{fail_vec}.}

\item{.generated_desc}{for convenience functions like
\link{sc_col_elements} to provide additional information about
the check.}
}
\value{
see \link{add_sanity_check}
}
\description{
NOTE the also add_sanity_check calls this function, the parameters
are documented in add_sanity_check because that function gets
exported.
}
