% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{h_collapse_char_vec}
\alias{h_collapse_char_vec}
\title{Collapse a vector of characters to a string with separators}
\usage{
h_collapse_char_vec(v, collapse = ", ", qoute = "'")
}
\arguments{
\item{v}{vector of chars to be collapsed}

\item{collapse}{character that separates the elements in the returned object}

\item{qoute}{character that surronds every element in \code{v} in the
returned object}
}
\value{
collapsed version of \code{v}
}
\description{
Collapse a vector of characters to a string with separators
}
\examples{
cat(sanityTracker:::h_collapse_char_vec(v = letters[1:4]))
}
