% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{h_deparsed_sys_call}
\alias{h_deparsed_sys_call}
\title{Simply converts a call into a character}
\usage{
h_deparsed_sys_call(which)
}
\arguments{
\item{which}{see \link{sys.call}. However the function bounds it by
the number of encolsing environments.}
}
\value{
the call of the corresponding environment as character
}
\description{
Simply converts a call into a character
}
